slots: bring in handling of parameter passing
Deciding which parameters can be passed where is very much compiler
logic, not dialog agent business
