+++ lib/ast/api.js
@@ -29,6 +29,7 @@ let { typeCheckFilter,
       typeCheckBookkeeping } = require('../typecheck');
 const {
     recursiveYieldArraySlots,
+    makeScope,
     InputParamSlot,
     FilterSlot,
     ArrayIndexSlot,
@@ -449,39 +450,6 @@ Ast.Table.prototype.iterateSlots2 = function* iterateSlots2(scope) {
         throw new TypeError("Can't handle " + this);
     }
 };
-
-
-function makeScope(invocation) {
-    // make out parameters available in the "scope", which puts
-    // them as possible options for a later slot fill
-    const schema = invocation.schema;
-    if (!schema)
-        return null;
-    const scope = {};
-    for (let argname in schema.out) {
-        let index = schema.index[argname];
-        let argcanonical = schema.argcanonicals[index] || argname;
-
-        let kind;
-        if (invocation.isVarRef)
-            kind = null;
-        else if (invocation.isResultRef)
-            kind = invocation.kind;
-        else
-            kind = invocation.selector.kind;
-        scope[argname] = {
-            value: Ast.Value.VarRef(argname),
-            type: schema.out[argname],
-            argcanonical: argcanonical,
-            kind: kind
-        };
-    }
-    scope['$event'] = {
-        value: Ast.Value.Event(null),
-        type: Type.String,
-    };
-    return scope;
-}
 
 Ast.Stream.prototype.iterateSlots = function* iterateSlotsStream(scope) {
     if (this.isVarRef) {
