+++ lib/ast/slots.js
@@ -14,10 +14,15 @@ const Type = require('../type');
 const I18n = require('../i18n');
 const { clean } = require('../utils');
 
+const { Value } = require('./values');
+
 class AbstractSlot {
     constructor(prim, scope) {
         this._prim = prim;
+
+
         this._scope = scope;
+        this._options = undefined;
     }
 
     get primitive() {
@@ -25,6 +30,23 @@ class AbstractSlot {
     }
     get scope() {
         return this._scope;
+    }
+
+    // the available options to parameter pass from
+    // this is computed lazily because it needs this.type, which
+    // is not available in the constructor
+    get options() {
+        if (this._options)
+            return this._options;
+
+        let options = [];
+        const slotType = this.type;
+        for (var vname in this._scope) {
+            let option = this._scope[vname];
+            if (Type.isAssignable(option.type, slotType))
+                options.push(option);
+        }
+        return this._options = options;
     }
 
     /* istanbul ignore next */
@@ -107,6 +129,27 @@ class FilterSlot extends AbstractSlot {
 
     get _argcanonical() {
         return this._arg ? this._arg.canonical : clean(this._filter.name);
+    }
+
+    // overidde the default option handling to filter out non-sensical filters such as "x == x"
+    get options() {
+        if (this._options)
+            return this._options;
+        let options = [];
+
+        const slotType = this.type;
+        for (var vname in this._scope) {
+            let option = this._scope[vname];
+            if (Type.isAssignable(option.type, slotType)) {
+                if (option.value.isVarRef && option.value.name === this._filter.name &&
+                    option._prim === this._prim)
+                    continue;
+                if (option.value.isEvent)
+                    continue;
+                options.push(option);
+            }
+        }
+        return this._options = options;
     }
 
     get type() {
@@ -331,6 +374,41 @@ class FieldSlot extends AbstractSlot {
     }
 }
 
+function makeScope(invocation) {
+    // make out parameters available in the "scope", which puts
+    // them as possible options for a later slot fill
+    const schema = invocation.schema;
+    if (!schema)
+        return null;
+    const scope = {};
+    for (let argname in schema.out) {
+        let index = schema.index[argname];
+        let argcanonical = schema.argcanonicals[index] || argname;
+
+        let kind;
+        if (invocation.isVarRef)
+            kind = null;
+        else if (invocation.isResultRef)
+            kind = invocation.kind;
+        else
+            kind = invocation.selector.kind;
+        scope[argname] = {
+            value: Value.VarRef(argname),
+            type: schema.out[argname],
+            argcanonical: argcanonical,
+
+            _prim: invocation,
+            kind: kind,
+            kind_canonical: schema.class ? (schema.class.metadata.canonical || null) : null,
+        };
+    }
+    scope['$event'] = {
+        value: Value.Event(null),
+        type: Type.String,
+    };
+    return scope;
+}
+
 function* recursiveYieldArraySlots(slot) {
     yield slot;
     const value = slot.get();
@@ -345,6 +423,7 @@ function* recursiveYieldArraySlots(slot) {
 
 module.exports = {
     recursiveYieldArraySlots,
+    makeScope,
     InputParamSlot,
     FilterSlot,
     ArrayIndexSlot,
